requirements.txt
asgiref==3.11.1
Django==6.0.3
django-simple-history==3.11.0
html5lib==1.1
pillow==12.1.1
python-decouple==3.8
pypdf==6.9.2
requests==2.32.3
reportlab==4.4.10
sqlparse==0.5.5
tzdata==2025.3
xhtml2pdf==0.2.17
